separate social post list to own components
