+++ src/components/Common/Social/PostComment/Item/index.js
@@ -3,8 +3,14 @@ import { Avatar, Button, Dropdown, Menu, Skeleton } from 'antd'
 import { MoreOutlined } from '@ant-design/icons'
 import { isNil } from 'lodash'
 import moment from 'moment'
+import { sendToSocialProfile } from 'components/utils'
+import { useHistory } from 'react-router-dom'
+import { useSelector } from 'react-redux'
 
 const PostCommentItem = ({ comment, isLoading }) => {
+  const user = useSelector(state => state.user)
+  const history = useHistory()
+
   const CommentMenu = () => {
     return (
       <Menu>
@@ -39,7 +45,13 @@ const PostCommentItem = ({ comment, isLoading }) => {
       <div className="col">
         <div className="row text-dark">
           <div className="col-12">
-            <span className="font-weight-bold">
+            <span
+              role="button"
+              tabIndex={0}
+              className="invisible-btn font-weight-bold"
+              onClick={() => sendToSocialProfile(history, user, comment.User?.accountId)}
+              onKeyDown={e => e.preventDefault()}
+            >
               {`${!isNil(comment.User?.firstName) ? comment.User?.firstName : 'Anonymous'} ${
                 !isNil(comment.User?.lastName) ? comment.User?.lastName : 'Pigeon'
               }`}
